Added proper start game deck builder option
reworked the gameloopmanager, to separate the leaving and entering of a state into a seperate structure, so that moving from one state to another is more flexible. used this rework to add a startgame mode and a build starting deck mode. also fixed a lot of bugs allong the way as they came up.